+++ pkg/api/annotations_test.go
@@ -14,12 +14,9 @@ import (
 	"github.com/grafana/grafana/pkg/api/response"
 	"github.com/grafana/grafana/pkg/api/routing"
 	"github.com/grafana/grafana/pkg/bus"
-	"github.com/grafana/grafana/pkg/components/simplejson"
 	"github.com/grafana/grafana/pkg/models"
 	"github.com/grafana/grafana/pkg/services/accesscontrol"
 	"github.com/grafana/grafana/pkg/services/annotations"
-	dashboardstore "github.com/grafana/grafana/pkg/services/dashboards/database"
-	"github.com/grafana/grafana/pkg/services/featuremgmt"
 	"github.com/grafana/grafana/pkg/services/guardian"
 	"github.com/grafana/grafana/pkg/services/sqlstore"
 	"github.com/grafana/grafana/pkg/services/sqlstore/mockstore"
@@ -956,140 +953,4 @@ func setUpFGACGuardian(t *testing.T) {
 	})
 
 	guardian.MockDashboardGuardian(&guardian.FakeDashboardGuardian{CanEditValue: true})
-}
-
-// Note: this is a database test, so it should be located in pkg/services/sqlstore/annotation_test.go
-// however, placing it there would currently result in circular dependencies.
-// So this test should be moved once the SQL package is separated out.
-func TestAnnotationListingWithFGAC(t *testing.T) {
-	sql := sqlstore.InitTestDB(t)
-	sql.Cfg.IsFeatureToggleEnabled = func(key string) bool {
-		return key == featuremgmt.FlagAccesscontrol
-	}
-	repo := sqlstore.NewSQLAnnotationRepo(sql)
-	dashboardStore := dashboardstore.ProvideDashboardStore(sql)
-
-	testDashboard1 := models.SaveDashboardCommand{
-		UserId: 1,
-		OrgId:  1,
-		Dashboard: simplejson.NewFromAny(map[string]interface{}{
-			"title": "Dashboard 1",
-		}),
-	}
-	dashboard, err := dashboardStore.SaveDashboard(testDashboard1)
-	require.NoError(t, err)
-	dash1UID := dashboard.Uid
-
-	testDashboard2 := models.SaveDashboardCommand{
-		UserId: 1,
-		OrgId:  1,
-		Dashboard: simplejson.NewFromAny(map[string]interface{}{
-			"title": "Dashboard 2",
-		}),
-	}
-	_, err = dashboardStore.SaveDashboard(testDashboard2)
-	require.NoError(t, err)
-
-	dash1Annotation := &annotations.Item{
-		OrgId:       1,
-		DashboardId: 1,
-		Epoch:       10,
-	}
-	err = repo.Save(dash1Annotation)
-	require.NoError(t, err)
-
-	dash2Annotation := &annotations.Item{
-		OrgId:       1,
-		DashboardId: 2,
-		Epoch:       10,
-	}
-	err = repo.Save(dash2Annotation)
-	require.NoError(t, err)
-
-	organizationAnnotation := &annotations.Item{
-		OrgId: 1,
-		Epoch: 10,
-	}
-	err = repo.Save(organizationAnnotation)
-	require.NoError(t, err)
-
-	user := &models.SignedInUser{
-		UserId: 1,
-		OrgId:  1,
-	}
-
-	type testStruct struct {
-		description           string
-		permissions           map[string][]string
-		expectedAnnotationIds []int64
-		expectedError         bool
-	}
-
-	testCases := []testStruct{
-		{
-			description: "Should find all annotations when has permissions to list all annotations and read all dashboards",
-			permissions: map[string][]string{
-				accesscontrol.ActionAnnotationsRead: {accesscontrol.ScopeAnnotationsAll},
-				accesscontrol.ActionDashboardsRead:  {accesscontrol.ScopeDashboardsAll},
-			},
-			expectedAnnotationIds: []int64{dash1Annotation.Id, dash2Annotation.Id, organizationAnnotation.Id},
-		},
-		{
-			description: "Should find all dashboard annotations",
-			permissions: map[string][]string{
-				accesscontrol.ActionAnnotationsRead: {accesscontrol.ScopeAnnotationsTypeDashboard},
-				accesscontrol.ActionDashboardsRead:  {accesscontrol.ScopeDashboardsAll},
-			},
-			expectedAnnotationIds: []int64{dash1Annotation.Id, dash2Annotation.Id},
-		},
-		{
-			description: "Should find only annotations from dashboards that user can read",
-			permissions: map[string][]string{
-				accesscontrol.ActionAnnotationsRead: {accesscontrol.ScopeAnnotationsTypeDashboard},
-				accesscontrol.ActionDashboardsRead:  {fmt.Sprintf("dashboards:uid:%s", dash1UID)},
-			},
-			expectedAnnotationIds: []int64{dash1Annotation.Id},
-		},
-		{
-			description: "Should find no annotations if user can't view dashboards or organization annotations",
-			permissions: map[string][]string{
-				accesscontrol.ActionAnnotationsRead: {accesscontrol.ScopeAnnotationsTypeDashboard},
-			},
-			expectedAnnotationIds: []int64{},
-		},
-		{
-			description: "Should find only organization annotations",
-			permissions: map[string][]string{
-				accesscontrol.ActionAnnotationsRead: {accesscontrol.ScopeAnnotationsTypeOrganization},
-				accesscontrol.ActionDashboardsRead:  {accesscontrol.ScopeDashboardsAll},
-			},
-			expectedAnnotationIds: []int64{organizationAnnotation.Id},
-		},
-		{
-			description: "Should error if user doesn't have annotation read permissions",
-			permissions: map[string][]string{
-				accesscontrol.ActionDashboardsRead: {accesscontrol.ScopeDashboardsAll},
-			},
-			expectedError: true,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.description, func(t *testing.T) {
-			user.Permissions = map[int64]map[string][]string{1: tc.permissions}
-			results, err := repo.Find(context.Background(), &annotations.ItemQuery{
-				OrgId:        1,
-				SignedInUser: user,
-			})
-			if tc.expectedError {
-				require.Error(t, err)
-				return
-			}
-			require.NoError(t, err)
-			assert.Len(t, results, len(tc.expectedAnnotationIds))
-			for _, r := range results {
-				assert.Contains(t, tc.expectedAnnotationIds, r.Id)
-			}
-		})
-	}
 }
